Move configs into single config script
