--- bin/collector-config.js
@@ -1,11 +1,0 @@
-// Default configuration for development.
-module.exports = {
-  "mongo-host": "127.0.0.1",
-  "mongo-port": 27017,
-  "mongo-database": "cube_development",
-  "mongo-username": null,
-  "mongo-password": null,
-  "http-port": 1080,
-  "udp-port": 1180,
-  "authenticator": "allow_all"
-};
